* Reorganise modules with more structure.   * Replace:     * Chairman.Base     * Chairman.Network     * Chairman.Process   * With:     * Chairman.Hedgehog.Base     * Chairman.Hedgehog.File     * Chairman.Hedgehog.Network     * Chairman.Hedgehog.Process     * Chairman.Monad     * Chairman.String     * Chairman.Time * Refactor `H.evalM . liftIO` to `H.evalIO` * Ensure that all the `note` functions annotate in the caller's context.  This ensures the annotations appear in the body of the tests and not in the annotation functions. * New `assertIsJsonFile`, replacing exist assertion that didn't really work. * New `lastLine` function for extracting the last line in a `String` * Remove -Wno-unused-imports -Wno-unused-local-binds -Wno-unused-matches
